+++ tests/test-graphs-for-known-bounds.R
@@ -97,7 +97,7 @@ bound <- optimize_effect(obj)
 obj <- analyze_graph(b, constraints = mono2, effectt = eff)
 bound <- optimize_effect(obj)
 
-## no idea if these are right or published 
+## compare to Arvids paper 2009 
 
 
 obj <- analyze_graph(b, constraints = list("Z(X=0)<=Z(X=1)"), effectt = eff)